requirements.txt
Django==5.2.8
djangorestframework
djangorestframework-simplejwt
django-graphql-jwt==0.4.0
django-graphql-auth
django-filter
graphene-django
django-redis
gql[all]
celery[redis]
django-celery-beat
django-cors-headers
pillow
python-decouple
drf-spectacular[sidecar]